// frontend/src/generated/models/marketRiskWorkspaceResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * CFO Command Center API
 * OpenAPI spec version: 0.3.0
 */
import type { MarketRiskWorkspaceResponseAssetsItem } from './marketRiskWorkspaceResponseAssetsItem';
import type { MarketRiskWorkspaceResponseAssurance } from './marketRiskWorkspaceResponseAssurance';
import type { WorkspaceContextResponse } from './workspaceContextResponse';
import type { MarketRiskWorkspaceResponseLineage } from './marketRiskWorkspaceResponseLineage';
import type { MarketRiskWorkspaceResponseSelectedRuns } from './marketRiskWorkspaceResponseSelectedRuns';
import type { MarketRiskWorkspaceResponseThresholdStatesItem } from './marketRiskWorkspaceResponseThresholdStatesItem';

export interface MarketRiskWorkspaceResponse {
  as_of: string;
  assets?: MarketRiskWorkspaceResponseAssetsItem[];
  assurance?: MarketRiskWorkspaceResponseAssurance;
  context: WorkspaceContextResponse;
  lineage?: MarketRiskWorkspaceResponseLineage;
  projection_version: number;
  selected_runs?: MarketRiskWorkspaceResponseSelectedRuns;
  source_snapshot_ids?: string[];
  threshold_states?: MarketRiskWorkspaceResponseThresholdStatesItem[];
}